WIP: Plugin: Fix dangling element pointer problem.
The pluginAPI layer uses wrapper objects to delegate
propery and method calls to the "real" MuseScore
objects that they represent. Prior to this commit,
if those "real" objects are deleted the wrappers
point to unallocated memory potentially causing crashes
if they are accessed or modified. This commit solves
this problem.